requirements.txt
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
Flask==2.1.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycparser==2.21
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.33
SQLAlchemy-Utils==0.38.2
Werkzeug==2.1.1
zipp==3.8.0
